Validate Glean pings from error stream (#33)
Because custom application ids (e.g. from dev builds) are not registered in probe-scraper, their pings do not pass validation at ingestion and end up in error stream. They have been shown in Debug Viewer with an "Unknown schema" error emitted in the decoder. This is correct from the point of view of the whole ingestion pipeline, but confusing to users of the Viewer application.

This adds an additional schema validation step for Glean pings from the error stream. If they pass it, they're no longer marked as errors in the UI but an icon with tooltip explaining additional validation step is shown instead.