When flattening don't treat `null` as an object
Despite our best efforts we do allow `null` as a value in e.g. object
metrics to be sent.
That previously crashed the frontend as it tried to dive into `null` as
if it was an object.
It's not, despite everything `typeof null` wants to tell you.
We handle that case explicitly now and just render `null` values as ...
nothing.

Fixes #233